app/backend/requirements.txt
# Core dependencies
redis>=5.0.0
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
pydantic>=2.6.0
websockets==12.0
anthropic==0.40.0
openai>=1.12.0
slowapi==0.1.9

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
